Drop tenant scoping from the storage mappings list query
Tenant scoping is moving into the auth token, so each query no longer carries
its own prefix filter. The server applies the caller's catalog-read scope on
its own, and the schema states that a filter can only narrow that scope, so an
absent filter cannot widen what a caller sees.

Remove `by: { underPrefix }` from `StorageMappingQuery`, along with the
`selectedTenant` reads in `useStorageMappings` and `useStorageMappingService`.
This also removes the last use of the deprecated `by` argument in this file.

The change repairs the edit dialog. The settings table lists every mapping the
caller can read, and a row click asks `UpdateMappingWizard` to resolve that
catalog prefix through this query. A tenant-scoped query held no row from
another prefix, so the dialog took its `notFound` path and closed itself.

`refetchMappings` now runs for every create and update. It previously returned
early whenever no tenant was selected.

Interim note: `PrefixCard` merges these prefixes with `useLiveSpecs`, which is
still scoped to the selected tenant, so its autocomplete draws the two sides
from different scopes until the token change lands.
