+++ src/gql-types/gql.ts
@@ -36,7 +36,7 @@ type Documents = {
     "\n    mutation CreateStorageMapping(\n        $catalogPrefix: Prefix!\n        $spec: JSON!\n        $detail: String\n    ) {\n        createStorageMapping(\n            catalogPrefix: $catalogPrefix\n            spec: $spec\n            detail: $detail\n        ) {\n            catalogPrefix\n        }\n    }\n": typeof types.CreateStorageMappingDocument,
     "\n    mutation UpdateStorageMapping(\n        $catalogPrefix: Prefix!\n        $spec: JSON!\n        $detail: String\n    ) {\n        updateStorageMapping(\n            catalogPrefix: $catalogPrefix\n            spec: $spec\n            detail: $detail\n        ) {\n            catalogPrefix\n            republish\n        }\n    }\n": typeof types.UpdateStorageMappingDocument,
     "\n    mutation TestConnectionHealth($catalogPrefix: Prefix!, $spec: JSON!) {\n        testConnectionHealth(catalogPrefix: $catalogPrefix, spec: $spec) {\n            results {\n                fragmentStore\n                dataPlaneName\n                error\n            }\n        }\n    }\n": typeof types.TestConnectionHealthDocument,
-    "\n    query StorageMappingQuery($underPrefix: Prefix!) {\n        storageMappings(by: { underPrefix: $underPrefix }) {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n": typeof types.StorageMappingQueryDocument,
+    "\n    query StorageMappingQuery {\n        storageMappings {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n": typeof types.StorageMappingQueryDocument,
     "\n    query StorageMappingsTable($first: Int, $after: String) {\n        storageMappings(first: $first, after: $after) {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n            pageInfo {\n                ...PageInfoFields\n            }\n        }\n    }\n": typeof types.StorageMappingsTableDocument,
     "\n    query AlertingOverviewQuery($prefix: String!, $active: Boolean) {\n        alerts(by: { prefix: $prefix, active: $active }) {\n            edges {\n                node {\n                    alertType\n                    firedAt\n                    catalogName\n                    alertDetails: arguments\n                    resolvedAt\n                }\n            }\n        }\n    }\n": typeof types.AlertingOverviewQueryDocument,
     "\n    query ActiveAlertCount($catalogName: Name!) {\n        liveSpecs(by: { names: [$catalogName] }) {\n            edges {\n                cursor\n                node {\n                    activeAlerts {\n                        alertType\n                    }\n                }\n            }\n        }\n    }\n": typeof types.ActiveAlertCountDocument,
@@ -68,7 +68,7 @@ const documents: Documents = {
     "\n    mutation CreateStorageMapping(\n        $catalogPrefix: Prefix!\n        $spec: JSON!\n        $detail: String\n    ) {\n        createStorageMapping(\n            catalogPrefix: $catalogPrefix\n            spec: $spec\n            detail: $detail\n        ) {\n            catalogPrefix\n        }\n    }\n": types.CreateStorageMappingDocument,
     "\n    mutation UpdateStorageMapping(\n        $catalogPrefix: Prefix!\n        $spec: JSON!\n        $detail: String\n    ) {\n        updateStorageMapping(\n            catalogPrefix: $catalogPrefix\n            spec: $spec\n            detail: $detail\n        ) {\n            catalogPrefix\n            republish\n        }\n    }\n": types.UpdateStorageMappingDocument,
     "\n    mutation TestConnectionHealth($catalogPrefix: Prefix!, $spec: JSON!) {\n        testConnectionHealth(catalogPrefix: $catalogPrefix, spec: $spec) {\n            results {\n                fragmentStore\n                dataPlaneName\n                error\n            }\n        }\n    }\n": types.TestConnectionHealthDocument,
-    "\n    query StorageMappingQuery($underPrefix: Prefix!) {\n        storageMappings(by: { underPrefix: $underPrefix }) {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n": types.StorageMappingQueryDocument,
+    "\n    query StorageMappingQuery {\n        storageMappings {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n": types.StorageMappingQueryDocument,
     "\n    query StorageMappingsTable($first: Int, $after: String) {\n        storageMappings(first: $first, after: $after) {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n            pageInfo {\n                ...PageInfoFields\n            }\n        }\n    }\n": types.StorageMappingsTableDocument,
     "\n    query AlertingOverviewQuery($prefix: String!, $active: Boolean) {\n        alerts(by: { prefix: $prefix, active: $active }) {\n            edges {\n                node {\n                    alertType\n                    firedAt\n                    catalogName\n                    alertDetails: arguments\n                    resolvedAt\n                }\n            }\n        }\n    }\n": types.AlertingOverviewQueryDocument,
     "\n    query ActiveAlertCount($catalogName: Name!) {\n        liveSpecs(by: { names: [$catalogName] }) {\n            edges {\n                cursor\n                node {\n                    activeAlerts {\n                        alertType\n                    }\n                }\n            }\n        }\n    }\n": types.ActiveAlertCountDocument,
@@ -183,7 +183,7 @@ export function graphql(source: "\n    mutation TestConnectionHealth($catalogPre
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n    query StorageMappingQuery($underPrefix: Prefix!) {\n        storageMappings(by: { underPrefix: $underPrefix }) {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n"): (typeof documents)["\n    query StorageMappingQuery($underPrefix: Prefix!) {\n        storageMappings(by: { underPrefix: $underPrefix }) {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n"];
+export function graphql(source: "\n    query StorageMappingQuery {\n        storageMappings {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n"): (typeof documents)["\n    query StorageMappingQuery {\n        storageMappings {\n            edges {\n                cursor\n                node {\n                    catalogPrefix\n                    spec\n                }\n            }\n        }\n    }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
